requirements.txt
annotated-types==0.7.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
flake8==7.3.0
idna==3.11
Jinja2==3.1.6
jsonschema==4.25.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
packaging==26.0
pycodestyle==2.14.0
pydantic==2.12.5
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.20.0
pyparsing==3.3.2
requests==2.33.0
rich==14.3.3
ruamel.yaml==0.19.1
ruamel.yaml.clib==0.2.15
shellingham==1.5.4
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
websocket-client==1.9.0
actions_toolkit==0.1.15
